// src/openapi/generated/models/paginated-response-task-application-daily-summary.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { TaskApplicationDailySummary } from '../models/task-application-daily-summary';
export interface PaginatedResponseTaskApplicationDailySummary {
  has_next: boolean;
  has_prev: boolean;
  items: Array<TaskApplicationDailySummary>;
  page: number;
  pages: number;
  size: number;
  total: number;
}