// src/openapi/generated/models/task-application-daily-summary.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */


/**
 * Esquema para resumen diario de aplicaciones de tareas por residente
 */
export interface TaskApplicationDailySummary {
  bed_name?: (string | null);
  date: string;
  first_application_time: string;
  last_application_time: string;
  resident_full_name: string;
  resident_id: string;
  task_count: number;
  task_types: Array<string>;
}